csic/service/energy/integration/requirements.txt
# FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
timescaledb>=2.13.0

# Redis
redis>=5.0.1
aioredis>=2.0.1

# Kafka
aiokafka>=0.10.0
kafka-python>=2.0.2

# ML and Data Processing
pandas>=2.1.0
numpy>=1.26.0
prophet>=1.1.5
scikit-learn>=1.4.0
scipy>=1.12.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Validation
email-validator>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
black>=23.0.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.2.0

# Monitoring
prometheus-client>=0.19.0
openTelemetry
